Adding validation for requiredBytesInCopy overflow
Also some uint32_t computations are now done on uint64_t.

Bug: dawn:482

+++ src/dawn_native/Queue.cpp
@@ -315,7 +315,7 @@ namespace dawn_native {
                 srcPointer += imageAdditionalStride;
             }
         } else {
-            uint64_t layerSize = rowsPerImageInBlock * actualBytesPerRow;
+            uint64_t layerSize = static_cast<uint64_t>(rowsPerImageInBlock) * actualBytesPerRow;
             if (!copyWholeData) {  // copy layer by layer
                 for (uint32_t d = 0; d < depth; ++d) {
                     memcpy(dstPointer, srcPointer, layerSize);
